--- browser/components/Dashboard/Viewport/Surveys/NewSurveyForm.scss
@@ -1,7 +1,0 @@
-.pt-select {
-	width: 40%;
-}
-
-.pt-intent-success {
-	float: right;
-}
