+++ browser/components/Dashboard/Viewport/Surveys/SurveySubcomponents/Table/Table.scss
