+++ browser/components/Dashboard/Viewport/Surveys/Surveys.jsx
@@ -1,17 +1,28 @@
 import React from 'react';
-import Survey from './Survey/SurveyContainer.js';
+// import Survey from './Survey/SurveyContainer.js';
 import './Surveys.scss';
-import Table from './SurveySubcomponents/Table.jsx';
+import Table from './SurveySubcomponents/Table/Table.jsx';
+import NewSurveyForm from './SurveySubcomponents/NewSurveyForm/NewSurveyForm.jsx'
+import ExistingSurvey from './SurveySubcomponents/ExistingSurveyEditor/ExistingSurveyEditor.jsx'
+
+import '../Viewport.scss';
+
+import { Overlay } from '@blueprintjs/core';
 
 export default ({
-  toggleForm,
+  toggleNewSurveyForm,
+  toggleExistingSurveyEditor,
   currentChannel,
+  modalType,
+  submitSurvey,
+  currentSurveyID,
+  showModal,
 }) => (
 	<div id="surveys" >
 
     <div className="pt-navbar survey-navbar">
       <div className="pt-navbar-group pt-align-right">
-        <button id='new-form-button' className='pt-button pt-intent-primary' onClick={toggleForm}>Create a New Survey</button>
+        <button id='new-form-button' className='pt-button pt-intent-primary' onClick={() => toggleNewSurveyForm()}>Create a New Survey</button>
 
         <div className="pt-input-group">
           <span className="pt-icon pt-icon-search"></span>
@@ -22,11 +33,36 @@ export default ({
     </div>
 
     <div className="pt-card">
-      <Table surveys={currentChannel.surveys}/>
+      <Table surveys={currentChannel.surveys} toggleExistingSurveyEditor={toggleExistingSurveyEditor}/>
     </div>
 
+    <Overlay 	className="pt-overlay-scroll-container"
+              isOpen={showModal}
+              inline={false}
+              canOutsideClickClose={true}
+              autoFocus={true}
+              enforceFocus={true}
+              canEscapeKeyClose={true}>
+      <div className="pt-card modal survey-modal">
+        {
+          modalType === 'new_survey'
+            ? <NewSurveyForm
+            toggleNewSurveyForm={toggleNewSurveyForm}
+            channel={currentChannel}
+            submitSurvey={submitSurvey} />
+            : null
+        }
+        {
+          modalType === 'existing_survey'
+            ? <ExistingSurveyEditor currentSurveyID={currentSurveyID} />
+            : null
+        }
+      </div>
+
+    </Overlay>
 
   </div>
+
 );
 
 //
