+++ browser/components/Dashboard/Viewport/Surveys/Surveys.scss
@@ -1,7 +1,4 @@
-h3 {
-	display: inline-block;
-}
-
-#new-form-button{
-	margin-right: 15px;
+.survey-modal {
+	width: 80%;
+	margin: 60px auto 0;
 }
