+++ browser/components/Dashboard/Viewport/Surveys/SurveysContainer.js
@@ -1,5 +1,5 @@
 import { connect } from 'react-redux';
-import SurveysComponent from './SurveysComponent';
+import SurveysComponent from './SurveysComponent.jsx';
 import { addSurveyToCurrent } from '../../../../reducers/channels';
 
 import axios from 'axios';
