+++ browser/index.html
@@ -3,13 +3,13 @@
 
 <head>
   <title>flow</title>
-  <link rel='stylesheet' href='https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css' />
+  <!--<link rel='stylesheet' href='https://maxcdn.bootstrapcdn.com/bootstrap/latest/css/bootstrap.min.css' />-->
 	<link rel='stylesheet' href='/css/blueprint.css' />
 	<link rel='stylesheet' href='/css/normalize.css' />
 
   <link rel='shortcut icon' href='/favicon.ico' />
   <link rel='stylesheet' href='/style.css' />
-  <link href='https://fonts.googleapis.com/css?family=Lato:300,400,700,900' rel='stylesheet'>
+  <!--<link href='https://fonts.googleapis.com/css?family=Lato:300,400,700,900' rel='stylesheet'>-->
 </head>
 
 <body>
